Thanks for the complements Casey's friends. End of Week Zero

+++ src/org/usfirst/frc/team78/robot/commands/DriveTime.java
@@ -11,7 +11,7 @@ public class DriveTime extends Command {
 	
 	double m_speed;
 	
-    public DriveTime(double speed, int time) {
+    public DriveTime(double speed, double time) {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
     	requires(Robot.chassis);
